Refactor annotations folder
