+++ src/summer/annotations/controller/Controller.java
@@ -1,11 +1,12 @@
-package src.summer.annotations;
+package src.summer.annotations.controller;
 
-import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import java.lang.annotation.ElementType;
 
 @Retention( RetentionPolicy.RUNTIME )
-@Target( ElementType.METHOD )
-public @interface Post {
+@Target( ElementType.TYPE )
+public @interface Controller {
+
 }
