+++ src/summer/annotations/controller/verb/Get.java
@@ -1,4 +1,4 @@
-package src.summer.annotations;
+package src.summer.annotations.controller.verb;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Retention( RetentionPolicy.RUNTIME )
 @Target( ElementType.METHOD )
-public @interface RestApi {
+public @interface Get {
 }
